Flag popup

+++ soen390/app/views/admin/narratives/flag.blade.php
@@ -1,30 +1,22 @@
-@extends('admin.master')
+@extends('admin.narratives.flagMaster')
 
 @section('view_title')
 Flagged Narrative(s)
 @stop
 
-@section('styles')
-<style>
-    th {
-        font-weight: 400;
-    }
-    .table-spinner td {
-        text-align: center;
-    }
-</style>
-@stop
-
 @section('content')
+<div class="narrative_remove text-center">
+    <label>Remove Narrative</label>&nbsp;<button type="button" class="btn btn-default" onclick="remove_narrative({{$NarrativeID}})"><i class="fa fa-trash-o fa-fw"></i></button>
+</div>
 <div class="message">
 </div>
 <table class="table narrative-table table-hover flag-table">
     <thead>
         <tr>
-            <th>#</th>
-            <th>{{ trans('admin.narratives.table.narrativeName') }}</th>
-            <th>{{ trans('admin.narratives.table.comment') }}</th>
-	        <th>{{ trans('admin.narratives.table.manage') }}</th>
+            <th style="width:100px;">#</th>
+            <th style="width:150px;">{{ trans('admin.narratives.table.narrativeName') }}</th>
+            <th style="width:750px;">{{ trans('admin.narratives.table.comment') }}</th>
+	        <th style="width:100px;">{{ trans('admin.narratives.table.manage') }}</th>
         </tr>
     </thead>
     <tbody class="table-spinner">
@@ -54,9 +46,7 @@ Flagged Narrative(s)
 		              	+ "<td>" + narrative.comment + "</td>"
                         + "<td>"
                         + "<div class=\"btn-group btn-group-xs\">"
-                        + "<button type=\"button\" class=\"btn btn-default\" onclick=\"remove_narrative("+ narrative.narrativeID+")\"><i class=\"fa fa-trash-o fa-fw\"></i></button>"
-                        + "<button type=\"button\" class=\"btn btn-default\" onclick=\"play_narrative("+ narrative.narrativeID +")\"><i class=\"fa fa-play fa-fw\"></i></button>"
-                        + "<button type=\"button\" class=\"btn btn-default\" onclick=\"remove_flag("+ narrative.id+")\"><i class=\"fa fa-times fa-fw\"></i></button>"
+                        + "<button type=\"button\" class=\"btn btn-default\" onclick=\"remove_flag("+ narrative.id+")\" data-toggle=\"tooltip\" data-placement=\"top\" title=\"Remove Flag\"><i class=\"fa fa-times fa-fw\"></i></button>"
                         + "</td></tr>");
                 });
 
@@ -85,28 +75,26 @@ Flagged Narrative(s)
     function remove_flag(id){//
         $.ajax({//
             type:'DELETE',
-            url:'flag/'+id,
+            url:'/admin/narrative/flag/'+id,
             success:function(data){//
                 $("tr#"+id).remove();
                 if($(".flag").length == 0){
-                    $(".flag-table").remove();
-                    $(".message").html("There are currently no reported narratives");
+                   alert("There are no more reports for this narrative. Window closing....");
+                   window.close();
                 }
             }
         });
     }
        function remove_narrative(id){//
-        $.ajax({//
-            type:'DELETE',
-            url:'narrative/'+id,
-            success:function(data){//
-               $("[data-NarrativeID="+id+"]").remove();
-                if($(".flag").length == 0){
-                    $(".flag-table").remove();
-                    $(".message").html("There are currently no reported narratives");
+        if(confirm("Are you sure you want to remove the entire narrative?")){
+            $.ajax({//
+                type:'DELETE',
+                url:'/admin/narrative/narrative/'+id,
+                success:function(data){//
+                    window.close();
                 }
-            }
-        });
+            });
+        }
     }
 </script>
